Implement ability to close kanalytics viewer app
This commit introduces the ability to close the kanalytics viewer app by introducing `disposeViewerAppWindow` and by using `DisposableEffect` on the `ViewerApp` composable. This is to ensure that we can properly dispose of the viewer app window when it's no longer needed.
Also added a new `ViewerAppViewController` for ios platform to host `ViewerApp` composable.