Implement showing json string as a tree and implement searching for key values
This commit updates several dependencies in `libs.versions.toml`, including adding `json-tree`.
It also removes `moko-permissions` and `moko-permissions-compose` dependencies as they are no longer needed.
The changes to the `SearchTextFieldAppBar` allow the user to customize the `placeholder` and add custom `actions`.

In `EventDetailsScreen`, JSON tree is added as a component and search is implemented. This screen now includes search functionality for JSON data, with next and previous arrows for search results.

Lastly, in `AnalyticsEventsScreen`, the `SearchTextFieldAppBar` component is updated to include a `placeholder`.

The color scheme and text styles for search results were also updated in `Color.kt` and `Theme.kt` files.

The sample app was also updated by removing the references to the removed dependencies.