Refactor KAnalyticsInterceptor and KAnalyticsCollector
This commit refactors the `KAnalyticsInterceptor` to directly take `numberOfDays` and `showNotification` as constructor parameters. The `KAnalyticsCollector` is updated to be initialized using `getInstance`, allowing for configuration of notification display and data retention days. Also, SampleViewModelFactory is changed to create KAnalytics using a function. Additionally `DayDuration` class is marked as JvmInline.