Refactor KAnalyticsInterceptor and KAnalyticsCollector
This commit refactors the `KAnalyticsInterceptor` to accept a `KAnalyticsCollector` instance in its constructor instead of `numberOfDays` and `showNotification`. It also removes unnecessary properties from `KAnalyticsCollector` and `RetentionPolicyManager` such as `repository`, `notificationManager` etc, as they're not used in the no-op artifact. `KAnalyticsCollector` no longer defaults to showing the notification by default, and the notification default is false. Finally, updates the `SampleViewModelFactory` to create the `KAnalytics` instance with the correct params.